fix(site): lift ConfigureAgentsDialog out of AgentCreateForm
ConfigureAgentsDialog was rendered inside AgentCreateForm, which is
only mounted when no agent is selected. This meant clicking the
sidebar settings button while viewing a conversation had no effect
because the dialog was not in the DOM.

Move the dialog to AgentsPageView so it is always mounted. Make
ConfigureAgentsDialog self-contained by internalizing the system
prompt and user prompt queries, mutations, and draft state that
previously lived in AgentCreateForm.

This also removes 4 props from AgentCreateForm and reduces the
ConfigureAgentsDialog interface from 15 props to 4.
